pyproject.toml
[tool.poetry]
name = "budget-bot-test"
version = "0.1.0"
description = ""
authors = ["Ivan <ishelukhin@tennisi.it>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^21.5"
python-dotenv = "^1.0.1"
aiosqlite = "^0.20.0"
google-oauth2-tool = "^0.0.3"
pandas = "^2.2.2"
gspread-asyncio = "^2.0.0"
pytz = "^2024.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
